Update out-of-place direct attack dialogue
I wrote these dialogues myself years ago (for Gideon, then Mercury was impressed and it travelled upstream) and left "Just {0}..." alone because I didn't know what value that variable took. Now I do, so I can edit it to something more tonally appropriate with the confident nature of the rest of the lines.