add "CreateGame" option to host rooms in the edopro lobby system
The option is a text string that represents a json object encoded in
base64, just like DbPaths.

Room options, as well as defaults, are in the CreateGameInfo class and
the json object is expected to match the class' schema.

In order for the json file to be considered non-empty (and consequently,
make windbot try to host a room instead of trying to join an existing
room), the "bestOf" option must be non-0.